refactor: TET-853 move FileItem component to the top level
