+++ src/components/FileItem/FileItem.props.ts
