+++ src/components/FileItem/FileItem.styles.ts
