+++ src/components/FileItem/stylesBuilder.ts
