+++ src/components/FileItem/types.ts
