+++ src/components/FileUploader/components/index.ts
@@ -1,4 +1,3 @@
 export * from './Control';
 export * from './DragAndDropField';
-export * from './FileItem';
 export * from './FilesList';
